+++ api/AltV.Net.Client/Elements/Entities/Vehicle.cs
@@ -18,7 +18,7 @@ namespace AltV.Net.Client.Elements.Entities
         public IntPtr VehicleNativePointer { get; private set; }
         public override IntPtr NativePointer => VehicleNativePointer;
 
-        public Vehicle(ICore core, IntPtr vehiclePointer, ushort id) : base(core, GetEntityPointer(core, vehiclePointer), id, BaseObjectType.Vehicle)
+        public Vehicle(ICore core, IntPtr vehiclePointer, uint id) : base(core, GetEntityPointer(core, vehiclePointer), id, BaseObjectType.Vehicle)
         {
             VehicleNativePointer = vehiclePointer;
         }
