+++ api/AltV.Net.Example/MyPlayerFactory.cs
@@ -6,7 +6,7 @@ namespace AltV.Net.Example
     [EntityFactory(BaseObjectType.Player)]
     public class MyPlayerFactory : IEntityFactory<IPlayer>
     {
-        public IPlayer Create(ICore core, IntPtr playerPointer, ushort id)
+        public IPlayer Create(ICore core, IntPtr playerPointer, uint id)
         {
             return new MyPlayer(core, playerPointer, id);
         }
